FIx doctest now that the correct sqrt functions are used

+++ nrpy/c_codegen.py
@@ -279,11 +279,11 @@ def c_codegen(
     ...     print(c_codegen(1/x**2 + 1/sp.sqrt(y) - 1/sp.sin(x*z), f"{fp_type} blah", include_braces=False, verbose=False, fp_type=fp_type))
     double blah = -1/sin(x*z) + (1.0/sqrt(y)) + (1.0/((x)*(x)));
     <BLANKLINE>
-    float blah = -1/sinf(x*z) + (1.0/sqrt(y)) + (1.0/((x)*(x)));
+    float blah = -1/sinf(x*z) + (1.0/sqrtf(y)) + (1.0/((x)*(x)));
     <BLANKLINE>
-    long double blah = -1/sinl(x*z) + (1.0/sqrt(y)) + (1.0/((x)*(x)));
+    long double blah = -1/sinl(x*z) + (1.0/sqrtl(y)) + (1.0/((x)*(x)));
     <BLANKLINE>
-    std::float32_t blah = -1/sinf(x*z) + (1.0/sqrt(y)) + (1.0/((x)*(x)));
+    std::float32_t blah = -1/sinf(x*z) + (1.0/sqrtf(y)) + (1.0/((x)*(x)));
     <BLANKLINE>
     std::float64_t blah = -1/sin(x*z) + (1.0/sqrt(y)) + (1.0/((x)*(x)));
     <BLANKLINE>
